powersave CPU policy
Suggested patch introduced powersaving cpu policies that trade off some
of the system's responsiveness for battery life. In my use case it
allows to go from ~2 to ~3.5 hours on the journey 100% -> 15%.

This feature only works when ampd is started with the -P flags.
Additionally, it can be mixed with auto mode by using both -A and -P
flags, which uses auto on AC power, and powersave on the battery.

This patch dynamically removes CPU cores when the system is idle to
prevent the scheduler from using them for any task. Additionally, it
puts halted CPU into the deepest supported C-state and bypass
latency-based logic to select target C-state i.e. assuming that it saves
save some power.

I also localize all CPU start and stop to a few functions, instead
setting bits.

It also introduce an upper limit on CPU frequency, which can be
controlled by the hw.powersavelimit at sysctl. I need this to prevent
the CPU from running in turbo mode, which consumes a lot of power. So
the default setting is 75, which is kind a guess, but which is good for
the i5-10210U which I'm using.

I have been using this patch for the couple of weeks and it works well.

Kernel + includes + sbin/sysctl + usr.sbin/apm{,d} recompilation required

Thus, this work is based on Solene's patch which she announced at
https://marc.info/?l=openbsd-tech&m=163259444331471&w=2

Co-authored-by: Solene Rapenne <solene@openbsd.org>